Update form Python2 to Python3. (Note old mbuild still uses python2)

+++ scripts/ipcthreadtrace.py
@@ -13,7 +13,7 @@ class IpcTrace:
     self.freq = sim.dvfs.get_frequency(0) # This script does not support DVFS
     args = dict(enumerate((args or '').split(':')))
     filename = args.get(0, None)
-    interval_ns = long(args.get(1, 10000))
+    interval_ns = int(args.get(1, 10000))
     if filename:
       self.fd = file(os.path.join(sim.config.output_dir, filename), 'w')
       self.isTerminal = False
